Add confidence bands to estimated Rt curves

Create an approximate confidence band for the Rt estimate. Note that the variance computation is approximate.

Usage

confband(object, lambda, level = 0.95, ...)

Arguments

object

a poisson_rt or cv_poisson_rt object.

lambda

the selected lambda. May be a scalar value, or in the case of cv_poisson_rt objects, "lambda.min" or "lambda.max".

level

the desired confidence level(s). These will be sorted if necessary.

...

additional arguments for methods. Unused.

Value

A data.frame containing the estimated Rt at the chosen lambda, and confidence limits corresponding to level

Examples

y <- c(1, rpois(100, dnorm(1:100, 50, 15)*500 + 1))
out <- estimate_rt(y, nsol = 10)
head(confband(out, out$lambda[2]))
#> An `rt_confidence_band` object.
#> 
#> * lambda = 55.638 
#> * degrees of freedom = 4 
#> 
#> # A tibble: 6 × 3
#>      Rt `2.5%` `97.5%`
#>   <dbl>  <dbl>   <dbl>
#> 1 0.629 0         2.00
#> 2 0.695 0         1.59
#> 3 0.763 0         1.65
#> 4 0.833 0         1.73
#> 5 0.905 0.0389    1.77
#> 6 0.978 0.116     1.84
head(confband(out, out$lambda[2], level = c(0.95, 0.8, 0.5)))
#> An `rt_confidence_band` object.
#> 
#> * lambda = 55.638 
#> * degrees of freedom = 4 
#> 
#> # A tibble: 6 × 7
#>      Rt `2.5%` `10.0%` `25.0%` `75.0%` `90.0%` `97.5%`
#>   <dbl>  <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#> 1 0.629 0        0       0.163    1.10    1.52    2.00
#> 2 0.695 0        0.110   0.388    1.00    1.28    1.59
#> 3 0.763 0        0.185   0.459    1.07    1.34    1.65
#> 4 0.833 0        0.249   0.527    1.14    1.42    1.73
#> 5 0.905 0.0389   0.342   0.610    1.20    1.47    1.77
#> 6 0.978 0.116    0.418   0.684    1.27    1.54    1.84

cv <- cv_estimate_rt(y, nfold = 3, nsol = 30)
head(confband(cv, "lambda.min", c(0.5, 0.9)))
#> An `rt_confidence_band` object.
#> 
#> * lambda = 0.371 
#> * degrees of freedom = 14 
#> 
#> # A tibble: 6 × 5
#>      Rt  `5%` `25%` `75%` `95%`
#>   <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 1.20      0 0.469  1.92  2.98
#> 2 0.991     0 0.466  1.52  2.28
#> 3 0.865     0 0.450  1.28  1.88
#> 4 0.792     0 0.425  1.16  1.69
#> 5 0.755     0 0.411  1.10  1.60
#> 6 0.746     0 0.411  1.08  1.57