Rt estimation algorithm configuration

We should convert this into an S3 method so that we can pass in a vector of counts or an rt_admm_configuration object (and overwrite)

Usage

configure_rt_admm(
  observed_counts,
  degree,
  weighted_past_counts = NULL,
  primal_var = NULL,
  auxi_var = NULL,
  dual_var = NULL,
  rho = -1,
  rho_adjust = -1,
  alpha = 0.5,
  gamma = 0.9,
  tolerance = 1e-04,
  maxiter_inner = 5L,
  maxiter_line = 5L,
  verbose = 0
)

Arguments

observed_counts

vector of the observed daily infection counts

degree

Integer. Degree of the piecewise polynomial curve to be estimated. For example, `degree = 0` corresponds to a piecewise constant curve.

weighted_past_counts

the weighted sum of past infections counts with corresponding serial interval functions (or its Gamma approximation) as weights

primal_var

initial values of log(Rt)

auxi_var

auxiliary variable in the ADMM algorithm

dual_var

dual variable in the ADMM algorithm

rho

Double. An ADMM parameter; coefficient of augmented term in the Lagrangian function.

rho_adjust

Double. An ADMM parameter; adjusted coefficient of augmented term in the Lagrangian function.

alpha

Double. A parameter adjusting upper bound in line search algorithm in `prox_newton` algorithm.

gamma

Double. A parameter adjusting step size in line search algorithm in `prox_newton` algorithm.

tolerance

Double. Tolerance of ADMM convergence.

maxiter_inner

Integer. Maximum number of iterations for the inner loop of proximal Newton method.

maxiter_line

Integer. Maximum number of iterations for the linesearch algorithm in the proximal Newton method.

verbose

Integer.

Value

a list of model parameters with class `rt_admm_configuration`