Add confidence bands to estimated Rt curves

Create an approximate confidence band for the Rt estimate. Note that the variance computation is approximate.

Usage

confband(object, lambda, level = 0.95, ...)

Arguments

object

a poisson_rt or cv_poisson_rt object.

lambda

the selected lambda. May be a scalar value, or in the case of cv_poisson_rt objects, "lambda.min" or "lambda.max".

level

the desired confidence level(s). These will be sorted if necessary.

...

additional arguments for methods. Unused.

Value

A data.frame containing the estimated Rt at the chosen lambda, and confidence limits corresponding to level

Examples

y <- c(1, rpois(100, dnorm(1:100, 50, 15) * 500 + 1))
out <- estimate_rt(y, nsol = 10)
head(confband(out, out$lambda[2]))
#> An `rt_confidence_band` object.
#> 
#> * lambda = 86.082 
#> * degrees of freedom = 4 
#> 
#> # A tibble: 6 × 3
#>      Rt `2.5%` `97.5%`
#>   <dbl>  <dbl>   <dbl>
#> 1 0.957 0         2.65
#> 2 0.989 0         2.03
#> 3 1.02  0.0554    1.99
#> 4 1.05  0.118     1.99
#> 5 1.08  0.210     1.96
#> 6 1.12  0.277     1.95
head(confband(out, out$lambda[2], level = c(0.95, 0.8, 0.5)))
#> An `rt_confidence_band` object.
#> 
#> * lambda = 86.082 
#> * degrees of freedom = 4 
#> 
#> # A tibble: 6 × 7
#>      Rt `2.5%` `10.0%` `25.0%` `75.0%` `90.0%` `97.5%`
#>   <dbl>  <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#> 1 0.957 0        0       0.378    1.53    2.06    2.65
#> 2 0.989 0        0.313   0.634    1.34    1.67    2.03
#> 3 1.02  0.0554   0.393   0.692    1.35    1.65    1.99
#> 4 1.05  0.118    0.446   0.734    1.37    1.66    1.99
#> 5 1.08  0.210    0.516   0.786    1.38    1.65    1.96
#> 6 1.12  0.277    0.570   0.829    1.40    1.66    1.95

cv <- cv_estimate_rt(y, nfold = 3, nsol = 30)
head(confband(cv, "lambda.min", c(0.5, 0.9)))
#> An `rt_confidence_band` object.
#> 
#> * lambda = 1.488 
#> * degrees of freedom = 10 
#> 
#> # A tibble: 6 × 5
#>      Rt   `5%` `25%` `75%` `95%`
#>   <dbl>  <dbl> <dbl> <dbl> <dbl>
#> 1  1.23 0      0.510  1.96  3.01
#> 2  1.16 0      0.665  1.66  2.38
#> 3  1.11 0.0603 0.682  1.54  2.16
#> 4  1.08 0.167  0.706  1.45  1.99
#> 5  1.06 0.167  0.696  1.42  1.95
#> 6  1.05 0.181  0.699  1.41  1.93