Leave-kth-out cross validation for choosing a optimal parameter lambda

Leave-kth-out cross validation for choosing a optimal parameter lambda

Usage

cv_estimate_rt(
  observed_counts,
  korder = 3L,
  dist_gamma = c(2.5, 2.5),
  nfold = 3L,
  error_measure = c("mse", "mae", "deviance"),
  x = 1:n,
  lambda = NULL,
  maxiter = 1000000L,
  delay_distn = NULL,
  ...
)

Arguments

observed_counts

vector of the observed daily infection counts

korder

Integer. Degree of the piecewise polynomial curve to be estimated. For example, korder = 0 corresponds to a piecewise constant curve.

dist_gamma

Vector of length 2. These are the shape and scale for the assumed serial interval distribution. Roughly, this distribution describes the probability of an infectious individual infecting someone else after some period of time after having become infectious. As in most literature, we assume that this interval follows a gamma distribution with some shape and scale.

nfold

Integer. This number of folds to conduct the leave-kth-out cross validation. For leave-kth-out cross validation, every kth observed_counts and their corresponding position (evenly or unevenly spaced) are placed into the same fold. The first and last observed_counts are not assigned to any folds. Smallest allowable value is nfold = 2.

error_measure

Metric used to calculate cross validation scores. Must be choose from mse, mae, and deviance. mse calculates the mean square error; mae calculates the mean absolute error; deviance calculates the deviance

x

a vector of positions at which the counts have been observed. In an ideal case, we would observe data at regular intervals (e.g. daily or weekly) but this may not always be the case. May be numeric or Date.

lambda

Vector. A user supplied sequence of tuning parameters which determines the balance between data fidelity and smoothness of the estimated Rt; larger lambda results in a smoother estimate. The default, NULL results in an automatic computation based on nlambda, the largest value of lambda that would result in a maximally smooth estimate, and lambda_min_ratio. Supplying a value of lambda overrides this behaviour. It is likely better to supply a decreasing sequence of lambda values than a single (small) value. If supplied, the user-defined lambda sequence is automatically sorted in decreasing order.

maxiter

Integer. Maximum number of iterations for the estimation algorithm.

delay_distn

in the case of a non-gamma delay distribution, a vector of delay probabilities may be passed here. These will be coerced to sum to 1, and padded with 0 in the right tail if necessary.

...

additional parameters passed to estimate_rt() function

Value

An object with S3 class "cv_poisson_rt". Among the list components:

full_fit An object with S3 class "poisson_rt", fitted with all observed_counts and lambda

cv_scores leave-kth-out cross validation scores

cv_se leave-kth-out cross validation standard error

lambda.min lambda which achieved the optimal cross validation score

lambda.1se lambda that gives the optimal cross validation score within one standard error.

lambda the value of lambda used in the algorithm.

Examples

y <- c(1, rpois(100, dnorm(1:100, 50, 15) * 500 + 1))
cv <- cv_estimate_rt(y, korder = 3, nfold = 3, nsol = 30)
cv
#> 
#> Call: cv_estimate_rt(observed_counts = y, korder = 3, nfold = 3, nsol = 30)
#> 
#> Degree of the estimated piecewise polynomial curve: 3 
#> 
#> Summary of cross validation across lambda:
#>                 lambda index cv_scores  cv_se dof
#> Min Lambda   378.19977     1     9.447 1.7235   4
#> CV Minimizer 145.85730     4     8.618 1.6177   4
#> 1se Lambda   378.19977     1     9.447 1.7235   4
#> Max Lambda     0.03782    30    11.136 0.5486  24
#>